Added Service Table Errors Updated Quota Datatypes Added DTOs related to Table Service Created rough Lock Service for Table Access Moved PreRemove Listeners to custom class to avoid singletons Changed some DTO Validation Checks to be uniform for Strings Create ServiceTableEntry Entity and Repo
Added Functionality to create Tables and get Columns from User DB in UserDbService

Laying Groundwork for ServiceTableEndpoint and implemented some of the endpoints

+++ goofy-protocol-fis-be/src/main/java/com/masl/goofy_protocol_fis_be/entity/UserQuotas.java
@@ -40,7 +40,7 @@ public class UserQuotas {
     @Column
     private Long tableMaxDbSize;
     @Column
-    private Integer tableMaxFieldSize;
+    private Long tableMaxFieldSize;
     @Column
     private Integer tableMaxTables;
     @Column
@@ -54,7 +54,7 @@ public class UserQuotas {
 
     // TableQuery
     @Column
-    private Integer tableQueryMaxQueryLength;
+    private Long tableQueryMaxQueryLength;
     @Column
     private Integer tableQueryMaxConditionCount;
     @Column
